requirements.txt
appnope==0.1.2
asgiref==3.4.1
auditlog3==1.0.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
django-multiselectfield==0.1.12
django-paypal==1.1.1
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-toolbelt==0.0.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
drf-dynamic-fields==0.3.1
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.9
inflection==0.5.1
install==1.3.4
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mixins==0.1.4
numpy==1.21.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==21.0
parso==0.8.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
pexpect==4.8.0
phonenumbers==8.12.27
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.2.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pybind11==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.3.0
pymongo==3.12.1
pyOpenSSL==20.0.1
pyparsing==3.0.1
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.10.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
static3==0.7.0
stripe==2.61.0
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.7.2
waitress==2.0.0
wcwidth==0.2.5
whitenoise==5.2.0
